+++ src/LazyAlgebra.jl
@@ -8,7 +8,7 @@
 # This file is part of LazyAlgebra (https://github.com/emmt/LazyAlgebra.jl)
 # released under the MIT "Expat" license.
 #
-# Copyright (c) 2017-2020 Éric Thiébaut.
+# Copyright (c) 2017-2021 Éric Thiébaut.
 #
 
 module LazyAlgebra
@@ -130,57 +130,7 @@ include("blas.jl")
 include("coder.jl")
 include("rules.jl")
 include("mappings.jl")
-
-"""
-    using LazyAlgebra.LazyAlgebraLowLevel
-
-imports types and methods that may be useful to extend or implement
-`LazyAlgebra` mappings.
-
-"""
-module LazyAlgebraLowLevel
-using ..LazyAlgebra
-for sym in (:Adjoint,
-            :AdjointInverse,
-            :DiagonalMapping,
-            :DiagonalType,
-            :Direct,
-            :Endomorphism,
-            :Inverse,
-            :InverseAdjoint,
-            :Linear,
-            :LinearType,
-            :Morphism,
-            :MorphismType,
-            :NonDiagonalMapping,
-            :NonLinear,
-            :NonSelfAdjoint,
-            :Operations,
-            :SelfAdjoint,
-            :SelfAdjointType,
-            :axpby_yields_zero,
-            :axpby_yields_y,
-            :axpby_yields_my,
-            :axpby_yields_by,
-            :axpby_yields_x,
-            :axpby_yields_xpy,
-            :axpby_yields_xmy,
-            :axpby_yields_xpby,
-            :axpby_yields_mx,
-            :axpby_yields_ymx,
-            :axpby_yields_mxmy,
-            :axpby_yields_bymx,
-            :axpby_yields_ax,
-            :axpby_yields_axpy,
-            :axpby_yields_axmy,
-            :axpby_yields_axpby,
-            :promote_multiplier)
-    @eval begin
-        import LazyAlgebra: $sym
-        export $sym
-    end
-end
-end # module LazyAlgebraLowLevel
+include("foundations.jl")
 
 include("sparse.jl")
 using .SparseOperators