+++ src/fft.jl
@@ -26,7 +26,7 @@ export
     ifftshift,
     rfftdims
 
-using ..LazyAlgebraLowLevel
+using ..Foundations
 using ..LazyAlgebra
 using ..LazyAlgebra:
     @callable, bad_argument, bad_size, compose