+++ test/fft-tests.jl
@@ -7,7 +7,7 @@ module TestingLazyAlgebraFFT
 
 using Test
 using LazyAlgebra
-using LazyAlgebra.LazyAlgebraLowLevel
+using LazyAlgebra.Foundations
 using AbstractFFTs, FFTW
 
 floats = (Float32, Float64)